requirement.txt
fastapi
uvicorn 
vapi_server_sdk
python-dotenv
